requirements.txt
numpy
scikit-learn
matplotlib
requests
python-dotenv
